+++ src/app/(home)/layout.tsx
@@ -1,6 +1,6 @@
 'use client';
-import Header from '@/components/common/Header';
-import MainContent from '@/components/common/MainContent';
+import Header from '@/components/common/header';
+import MainContent from '@/components/common/main-content';
 import FirstSidebar from '@/components/common/FirstSidebar';
 import React from 'react';
