+++ src/components/common/main-content.tsx
