+++ src/components/dashboards/StudyTypes/AddStudyTypeButton.tsx
@@ -1,4 +1,4 @@
-import Modal from '@/components/common/Modal';
+import Modal from '@/components/common/modal';
 import StudyTypeForm from '@/components/forms/StudyTypeForm';
 import useModal from '@/hooks/useModal';
 import AddBoxIcon from '@mui/icons-material/AddBox';
