+++ src/components/dashboards/StudyTypes/EditStudyTypeButton.tsx
@@ -1,4 +1,4 @@
-import Modal from '@/components/common/Modal';
+import Modal from '@/components/common/modal';
 import StudyTypeForm from '@/components/forms/StudyTypeForm';
 import useModal from '@/hooks/useModal';
 import EditIcon from '@mui/icons-material/Edit';
