+++ src/components/forms/DiseaseForm.tsx
@@ -2,7 +2,7 @@ import { useForm } from 'react-hook-form';
 import { Disease } from '@prisma/client';
 import { FieldConfig } from '@/types/FieldConfig';
 import { useSubmitForm } from '@/hooks/useSubmitForm';
-import NewForm from '../common/NewForm';
+import NewForm from '../common/new-form';
 
 interface Props {
   oldDisease?: Disease;
