+++ src/components/forms/PasswordChangeForm.tsx
@@ -2,7 +2,7 @@
 import { Button, FormControl, FormLabel, Input } from '@mui/joy';
 import { useState } from 'react';
 import { useForm } from 'react-hook-form';
-import Container from '../common/Container';
+import Container from '../common/container';
 import { fetchData } from '@/utils/fetchData';
 
 export default function PasswordChangeForm() {
