+++ src/components/forms/PatientForm.tsx
@@ -1,9 +1,9 @@
 import { Select, Option, FormControl, FormLabel } from '@mui/joy';
 import { useState } from 'react';
 import { useForm } from 'react-hook-form';
-import SubmitButton from '../common/SubmitButton';
+import SubmitButton from '../common/submit-button';
 import { FieldConfig } from '@/types/FieldConfig';
-import FormFieldsMapper from '../common/FormFieldsMapper';
+import FormFieldsMapper from '../common/form-fields-mapper';
 import { useSubmitForm } from '@/hooks/useSubmitForm';
 import { FullPatient } from '@/types/FullPatient';
