+++ src/components/forms/TreatmentTypeResultForm.tsx
@@ -1,7 +1,7 @@
 import { useForm } from 'react-hook-form';
 import { TreatmentTypeResult } from '@prisma/client';
 import { FieldConfig } from '@/types/FieldConfig';
-import Form from '../common/Form';
+import Form from '../common/form';
 import { useSubmitForm } from '@/hooks/useSubmitForm';
 
 interface Props {
